better radio look, clean up
