+++ sass/components/all.scss
@@ -1,5 +1,6 @@
 @import 'breadcrumbs';
 @import 'charts';
+@import 'forms';
 @import 'glossary';
 @import 'layout';
 @import 'progress';
